tetra: generate: move it to a different package
No functional changes.
